Fix Code formatting and decoupling
