+++ Dialer/Dialer/Utilities/ContactsFetcher.swift
@@ -10,7 +10,7 @@ import Contacts
 
 class PhoneContacts {
     
-    class func getContacts(filter: ContactsFilter = .none) -> [CNContact] {
+    class private func getContacts(filter: ContactsFilter = .none) -> [CNContact] {
         
         let contactStore = CNContactStore()
         let keysToFetch = [
@@ -39,5 +39,23 @@ class PhoneContacts {
             }
         }
         return results
+    }
+    
+    class func getMtnContacts() -> [Contact] {
+        var resultingContacts: [Contact] = []
+        let contacts = PhoneContacts.getContacts()
+        for contact in contacts {
+            if contact.phoneNumbers.count > 0  {
+                let contactPhoneNumbers = contact.phoneNumbers
+                let mtnNumbers = contactPhoneNumbers.filter { $0.value.stringValue.isMtnNumber }
+                
+                let numbers = mtnNumbers.compactMap { $0.value.value(forKey: "digits") as? String }
+                if mtnNumbers.isEmpty == false {
+                    let newContact = Contact(names:contact.givenName + " " +  contact.familyName,phoneNumbers: numbers)
+                    resultingContacts.append(newContact)
+                }
+            }
+        }
+        return resultingContacts
     }
 }
